@@ -1,4 +1,4 @@
-import { Grid, Typography } from "@material-ui/core"
+import { Grid, Typography } from "@mui/material"
 
 export const AttributionSponsors = () => {
   const sponsors = [
