@@ -1,8 +1,8 @@
-import { createTheme } from '@material-ui/core/styles';
-import { red } from '@material-ui/core/colors';
+import { createTheme, adaptV4Theme } from '@mui/material/styles';
+import { red } from '@mui/material/colors';
 
 // Create a theme instance.
-const theme = createTheme({
+const theme = createTheme(adaptV4Theme({
   palette: {
     primary: {
       main: '#0D78D3',
@@ -17,6 +17,6 @@ const theme = createTheme({
       default: '#fff',
     },
   },
-});
+}));
 
 export default theme;
